Add webpage to backend
BEFORE, development of the website took place independently of the
actual backend of the project. This meant that the front end was
not part of the actual backend structure, and stuck as static.

AFTER, the HTML used in the website has been added as an editable
template, and the site is now served and available alongside the API
and the rest of the backend.