Remove menu drop downs, Edit access to links/dir
For this, I simply cleared any sub-elements from the menu, such that
it is now a horizontal row with elements that do not collapse. I then
changed the link reference for the "Home" and "Contact us" links to
a simple "/" just to remove errors for now. Also, I changed the file
path for the images and the js files to the root folder above that of
the webpage file (public folder) to clear up some errors in the image
and js file accessing.