Complete API black box model
BEFORE, the API wasn't exactly usable, as it would not return correct
output when interacted with from the outside world.

AFTER, the black box model of the API is complete, in that while it may
not correctly perform the needed computations to compute the correct
results, it appears to do so, with output matching spec.